requirements.txt
PIL
convertapi
oauth2client
gspread
python-telegram-bot==11.1.0
Pillow